Break up WebPack config so both the build and Storybook are happy.
